Fix and reformatting

+++ contracts/strategies/CompoundStrategy.sol
@@ -13,18 +13,38 @@ interface IFactory {
 }
 
 interface IPair {
-    function totalSupply() external view returns (uint);
-    function balanceOf(address owner) external view returns (uint);
+    function totalSupply() external view returns (uint256);
+
+    function balanceOf(address owner) external view returns (uint256);
+
     function token0() external view returns (address);
+
     function token1() external view returns (address);
-    function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast);
-    function swap(uint amount0Out, uint amount1Out, address to, bytes calldata data) external;
+
+    function getReserves()
+        external
+        view
+        returns (
+            uint112 reserve0,
+            uint112 reserve1,
+            uint32 blockTimestampLast
+        );
+
+    function swap(
+        uint256 amount0Out,
+        uint256 amount1Out,
+        address to,
+        bytes calldata data
+    ) external;
 }
 
 interface IcToken is IERC20 {
-    function mint(uint mintAmount) external returns (uint256);
-    function redeem(uint redeemTokens) external returns (uint256);
+    function mint(uint256 mintAmount) external returns (uint256);
+
+    function redeem(uint256 redeemTokens) external returns (uint256);
+
     function redeemUnderlying(uint256 redeemAmount) external returns (uint256);
+
     function balanceOfUnderlying(address account) external returns (uint256);
 }
 
@@ -39,20 +59,27 @@ contract CompoundStrategy is IStrategy, BoringOwnable {
     IERC20 public immutable compToken;
     IERC20 public immutable weth;
     IFactory public immutable factory;
-    uint256 public impactDivisor;  
+    uint256 public impactDivisor;
     bool public exited;
 
-    constructor(address bentobox_, IFactory factory_, IERC20 token_, IcToken cToken_, IERC20 compToken_, IERC20 weth_) public {
+    constructor(
+        address bentobox_,
+        IFactory factory_,
+        IERC20 token_,
+        IcToken cToken_,
+        IERC20 compToken_,
+        IERC20 weth_
+    ) public {
         bentobox = bentobox_;
         factory = factory_;
         token = token_;
         cToken = cToken_;
         compToken = compToken_;
         weth = weth_;
-        impactDivisor = 10;    
+        impactDivisor = 10;
     }
 
-    function setImpactDivisor (uint256 impactDivisor_) public onlyOwner{
+    function setImpactDivisor(uint256 impactDivisor_) public onlyOwner {
         impactDivisor = impactDivisor_;
     }
 
@@ -74,21 +101,17 @@ contract CompoundStrategy is IStrategy, BoringOwnable {
         (uint256 reserve0, uint256 reserve1, ) = pair.getReserves();
         uint256 amountInWithFee = amountIn.mul(997);
         if (fromToken == pair.token0()) {
-            amountOut =
-                amountIn.mul(997).mul(reserve1) /
-                reserve0.mul(1000).add(amountInWithFee);
+            amountOut = amountIn.mul(997).mul(reserve1) / reserve0.mul(1000).add(amountInWithFee);
             IERC20(fromToken).safeTransfer(address(pair), amountIn);
             pair.swap(0, amountOut, to, new bytes(0));
             require(amountIn < reserve0 / impactDivisor, "CompoundStrategy: Impact");
         } else {
-            amountOut =
-                amountIn.mul(997).mul(reserve0) /
-                reserve1.mul(1000).add(amountInWithFee);
+            amountOut = amountIn.mul(997).mul(reserve0) / reserve1.mul(1000).add(amountInWithFee);
             IERC20(fromToken).safeTransfer(address(pair), amountIn);
             pair.swap(amountOut, 0, to, new bytes(0));
             require(amountIn < reserve1 / impactDivisor, "CompoundStrategy: Impact");
         }
-    }    
+    }
 
     // Send the assets to the Strategy and call skim to invest them
     /// @inheritdoc IStrategy
@@ -102,7 +125,7 @@ contract CompoundStrategy is IStrategy, BoringOwnable {
     function harvest(uint256 balance, address sender) external override onlyBentobox returns (int256 amountAdded) {
         // Get the amount of tokens that the cTokens currently represent
         uint256 tokenBalance = cToken.balanceOfUnderlying(address(this));
-        // Find out how much has been added from compounding interest. 
+        // Find out how much has been added from compounding interest.
         // If it's negative due to rounding (near impossible), just revert. Should be positive soon enough.
         uint256 amountAdded_ = tokenBalance.sub(balance);
         // Convert enough cToken to take out the profit
@@ -158,10 +181,14 @@ contract CompoundStrategy is IStrategy, BoringOwnable {
         exited = true;
     }
 
-    function afterExit(address to, uint256 value, bytes memory data) public onlyOwner returns(bool success) {
-        // After exited, the owner can perform ANY call. This is to rescue any funds that didn't get released during exit or 
+    function afterExit(
+        address to,
+        uint256 value,
+        bytes memory data
+    ) public onlyOwner returns (bool success) {
+        // After exited, the owner can perform ANY call. This is to rescue any funds that didn't get released during exit or
         // got earned afterwards due to vesting or airdrops, etc.
         require(exited, "CompoundStrategy: Not exited");
-        (success,) = to.call{ value: value }(data);
+        (success, ) = to.call{value: value}(data);
     }
 }
